moved progressbar to separate component
