+++ src/components/Tool/Appeal/Appeal.tsx
@@ -1,16 +1,10 @@
-import {
-  Box,
-  CardContent,
-  Typography,
-  Tooltip,
-  makeStyles,
-  Theme,
-} from '@material-ui/core';
+import { Box, CardContent, Typography, makeStyles } from '@material-ui/core';
 import React, { ReactElement } from 'react';
 import StarIcon from '@material-ui/icons/Star';
 import StarOutlineIcon from '@material-ui/icons/StarOutline';
 import theme from '../../../../src/theme';
 import AnimatedCard from '../../../../src/components/AnimatedCard';
+import AppealProgressBar from './AppealProgressBar';
 
 export interface Props {
   name: string;
@@ -23,7 +17,7 @@ export interface Props {
   total: number;
 }
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles(() => ({
   cardContent: {
     marginTop: -theme.spacing(2),
   },
@@ -89,107 +83,13 @@ const Appeal = ({
               )}
             </Box>
           </Box>
-          <Box display="flex" justifyContent="end">
-            <Tooltip title="Given" placement="top" arrow>
-              <Typography
-                variant="body2"
-                display="inline"
-                className={classes.colorYellow}
-              >
-                {given} {amountCurrency} (
-                {`${((given / (amount || 1)) * 100).toFixed(0)}%`})
-              </Typography>
-            </Tooltip>
-            <Typography
-              variant="body2"
-              display="inline"
-              style={{
-                padding: '0 3px 0 3px',
-              }}
-            >
-              /
-            </Typography>
-            <Tooltip title="Received" placement="top" arrow>
-              <Typography
-                variant="body2"
-                display="inline"
-                className={classes.colorOrange}
-              >
-                {received + given} {amountCurrency} (
-                {`${(((received + given) / (amount || 1)) * 100).toFixed(0)}%`})
-              </Typography>
-            </Tooltip>
-            <Typography
-              variant="body2"
-              display="inline"
-              style={{
-                padding: '0 3px 0 3px',
-              }}
-            >
-              /
-            </Typography>
-            <Tooltip title="Commited" placement="top" arrow>
-              <Typography
-                variant="body2"
-                display="inline"
-                className={classes.colorLightGray}
-              >
-                {commited + received + given} {amountCurrency} (
-                {`${(
-                  ((commited + received + given) / (amount || 1)) *
-                  100
-                ).toFixed(0)}%`}
-                )
-              </Typography>
-            </Tooltip>
-          </Box>
-          <Box className={classes.progressBarOuter}>
-            <Box
-              display="flex"
-              justifyContent="flex-start"
-              className={classes.progressBarInner}
-            >
-              <Tooltip title="Given" placement="top-start" arrow>
-                <Box
-                  style={{
-                    minWidth: `min(${(given / (amount || 1)) * 100}%,100%)`,
-                    height: '100%',
-                    backgroundColor: theme.palette.progressBarYellow.main,
-                    borderTopLeftRadius: 8,
-                    borderBottomLeftRadius: 8,
-                    borderTopRightRadius: !received && !commited ? 8 : 0,
-                    borderBottomRightRadius: !received && !commited ? 8 : 0,
-                  }}
-                />
-              </Tooltip>
-              <Tooltip title="Received" placement="top-start" arrow>
-                <Box
-                  style={{
-                    minWidth: `${(received / (amount || 1)) * 100}%`,
-                    height: '100%',
-                    backgroundColor: theme.palette.progressBarOrange.main,
-                    borderTopLeftRadius: !given ? 8 : 0,
-                    borderBottomLeftRadius: !given ? 8 : 0,
-                    borderTopRightRadius: !commited ? 8 : 0,
-                    borderBottomRightRadius: !commited ? 8 : 0,
-                  }}
-                />
-              </Tooltip>
-              <Tooltip title="Commited" placement="top-start" arrow>
-                <Box
-                  style={{
-                    minWidth: `${(commited / (amount || 1)) * 100}%`,
-                    height: '100%',
-                    backgroundColor: theme.palette.progressBarGray.main,
-                    borderTopLeftRadius: !given && !received ? 8 : 0,
-                    borderBottomLeftRadius: !given && !received ? 8 : 0,
-                    borderTopRightRadius: 8,
-                    borderBottomRightRadius: 8,
-                  }}
-                ></Box>
-              </Tooltip>
-            </Box>
-          </Box>
+          <AppealProgressBar
+            given={given}
+            commited={commited}
+            received={received}
+            amount={amount}
+            amountCurrency={amountCurrency}
+          />
         </CardContent>
       </AnimatedCard>
     </Box>
